Fix race condition in EventBus register/unregister on the same address - fixes #1531
The ConcurrentMap updates has been modified to use a compute operation in which the Handlers list is copied and updated. The compute
operation ensures that there is no modification race, the Handlers list modification is side effect free and the copied list is safely published
using final HandlerHolder array in the constructor.

Since the HandlerHolder is not anymore accessible, the HandlerRegistration has been refactored so the HandlerHolder reference is kept
instead on the HandlerRegistration where it replaces the registered boolean field instead. So now unregistration is done
by providing the HandlerHolder instead of the HandlerRegistration reference.

A test reproducing the issue has been added, the test performs intensive concurrency handler registration/unregistration of the same
address by two threads